Feature/summary update

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -1,7 +1,7 @@
 -- пользователи
 create table if not exists users
 (
-    id   serial primary key,
+    id   bigserial primary key,
     name varchar(255) not null
 );
 
@@ -16,14 +16,14 @@ create table if not exists registered_users
 -- Незарегистрированные пользователи (по сути просто ники людей, которые не зарегистрированы)
 create table if not exists not_registered_users
 (
-    user_id bigint primary key,
+            user_id bigint primary key,
     foreign key (user_id) references users (id)
 );
 
 -- Ивент
 create table if not exists events
 (
-    id          serial primary key,
+    id          bigserial primary key,
     name        varchar(255)  not null,
     description varchar(2055) not null
 );
@@ -31,7 +31,7 @@ create table if not exists events
 -- Чеки
 create table if not exists receipts
 (
-    id              serial primary key,
+    id              bigserial primary key,
     name            varchar(255) not null,
     event_id        bigint       not null,
     paid_by_user_id bigint default null,
@@ -42,7 +42,7 @@ create table if not exists receipts
 -- Расходы в чеке
 create table if not exists expenses
 (
-    id         serial primary key,
+    id         bigserial primary key,
     name       varchar(255)   not null,
     amount     decimal(19, 2) not null,
     receipt_id bigint         not null,
@@ -52,7 +52,7 @@ create table if not exists expenses
 -- Участники расхода
 create table if not exists expense_members
 (
-    id         serial primary key,
+    id         bigserial primary key,
     user_id    bigint default null,
     expense_id bigint not null,
     foreign key (user_id) references users (id),
